requirements.txt
streamlit==1.45.1
easyocr==1.7.2
pdf2image==1.17.0
Pillow==11.2.1
opencv-python-headless==4.11.0.86
scikit-learn==1.6.1
xgboost==2.1.4
numpy==1.26.4
